refactor parameters and results
