+++ json-rpc/src/main/scala/com/github/apalachemc/apalache/jsonrpc/JsonParameterParser.scala
@@ -4,21 +4,6 @@ import com.fasterxml.jackson.core.TreeNode
 import com.fasterxml.jackson.databind.ObjectMapper
 
 /**
- * Parameters for loading a specification in the JSON-RPC server.
- * @param sources
- *   A sequence of specification modules, each encoded in base64. The head must be the root module, and the rest are
- *   additional modules that are imported by the root module (except the standard ones).
- */
-case class LoadSpecParams(sources: Seq[String])
-
-/**
- * Parameters for disposing of a session in the JSON-RPC server.
- * @param sessionId
- *   the ID of the session to be disposed. The session ID is generated by the server when a specification is loaded.
- */
-case class DisposeSpecParams(sessionId: String)
-
-/**
  * A parser for JSON parameters used in the JSON-RPC server.
  * @param mapper
  *   An instance of ObjectMapper for converting JSON values.
