+++ json-rpc/src/main/scala/com/github/apalachemc/apalache/jsonrpc/JsonRpcServer.scala
@@ -23,132 +23,6 @@ import scala.collection.immutable
 import scala.util.Try
 
 /**
- * All kinds of the results that the exploration service can return.
- */
-sealed abstract class ExplorationServiceResult
-
-/**
- * Parameters for preparing a symbolic transition in the solver context.
- * @param sessionId
- *   the ID of the previously loaded specification
- * @param snapshotId
- *   the snapshot ID for recovering the context before the transition is assumed.
- *   If it is negative, no snapshot recovery is performed.
- * @param transitionId
- *   the number of transition to prepare, starting from 0. On step 0, it must be in the range `[0, nInitTransitions)`,
- *   on step 1 and later, it must be in the range `[0, nNextTransitions)`.
- * @param checkEnabled
- *   whether to check if the transition is enabled. If `false`, the transition is prepared and assumed, but no
- *   satisfiability is checked
- * @param timeoutSec
- *   the timeout in seconds for checking satisfiability. If `0`, the default timeout is used. This parameter is ignored
- *   if `checkEnabled` is `false`.
- */
-case class AssumeTransitionParams(
-    sessionId: String,
-    snapshotId: Int = -1,
-    transitionId: Int,
-    checkEnabled: Boolean,
-    timeoutSec: Int = 0)
-
-object AssumeTransitionParams {
-  def apply(sessionId: String, snapshotId: Int, transitionId: Int): AssumeTransitionParams = {
-    new AssumeTransitionParams(sessionId, snapshotId, transitionId, checkEnabled = true, timeoutSec = 0)
-  }
-
-  def apply(sessionId: String, snapshotId: Int, transitionId: Int, checkEnabled: Boolean): AssumeTransitionParams = {
-    new AssumeTransitionParams(sessionId, snapshotId, transitionId, checkEnabled, timeoutSec = 0)
-  }
-}
-
-object TransitionStatus {
-  type T = String
-  val ENABLED = "ENABLED"
-  val DISABLED = "DISABLED"
-  val UNKNOWN = "UNKNOWN"
-}
-
-/**
- * The result of preparing a symbolic transition.
- * @param sessionId
- *   the ID of the previously loaded specification
- * @param snapshotId
- *   the snapshot ID for recovering the context after the transition has been assumed.
- * @param transitionId
- *   the number of the prepared transition
- * @param status
- *   status of the transition: "ENABLED", "DISABLED", or "UNKNOWN"
- */
-case class AssumeTransitionResult(
-    sessionId: String,
-    snapshotId: Int,
-    transitionId: Int,
-    status: TransitionStatus.T)
-    extends ExplorationServiceResult
-
-/**
- * The result of loading a specification.
- * @param sessionId
- *   the new session ID
- * @param snapshotId
- *   the snapshot ID for recovering the context right after loading the specification.
- * @param specParameters
- *   the specification parameters that are needed for symbolic path exploration
- */
-case class LoadSpecResult(sessionId: String, snapshotId: Int, specParameters: SpecParameters)
-    extends ExplorationServiceResult
-
-/**
- * Specification parameters that are needed for symbolic path exploration. These numbers may be different from what the
- * user expects by reading the specification, as transitions and invariants are decomposed.
- *
- * @param nInitTransitions
- *   the number of initial symbolic transitions
- * @param nNextTransitions
- *   the number of next symbolic transitions
- * @param nStateInvariants
- *   the number of state invariants
- * @param nActionInvariants
- *   the number of action invariants
- * @param nTraceInvariants
- *   the number of trace invariants
- * @param hasView
- *   whether a view predicate is present in the specification
- */
-case class SpecParameters(
-    nInitTransitions: Int,
-    nNextTransitions: Int,
-    nStateInvariants: Int,
-    nActionInvariants: Int,
-    nTraceInvariants: Int,
-    hasView: Boolean)
-
-/**
- * The result of disposing a specification.
- * @param sessionId
- *   the ID of the previously loaded specification
- */
-case class DisposeSpecResult(sessionId: String) extends ExplorationServiceResult
-
-/**
- * Parameters for switching to the next step in symbolic path exploration.
- * @param sessionId
- *   the ID of the previously loaded specification
- */
-case class NextStepParams(sessionId: String)
-
-/**
- * The result of switching to the next step in symbolic path exploration.
- * @param sessionId
- *   the ID of the previously loaded specification
- * @param snapshotId
- *   the snapshot ID for recovering the context right after loading the specification.
- * @param newStepNo
- *   the number of the new step
- */
-case class NextStepResult(sessionId: String, snapshotId: Int, newStepNo: Int) extends ExplorationServiceResult
-
-/**
  * An error that can occur in the exploration service.
  * @param code
  *   error code, typically a JSON-RPC error code
